requirements.txt
# Core Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0

# Essential Data Processing  
numpy>=1.26.1,<2.0.0  # Pin to avoid NumPy 2.x issues
pillow>=11.0.0
h5py>=3.7.0

# Training & ML Libraries
scikit-learn>=1.0.2    # For TSNE in visualization
transformers>=4.21.0   # For learning rate scheduler
matplotlib>=3.3.0      # For plots and visualization
seaborn>=0.12.0        # Enhanced plotting (if used)

# Configuration & Utilities
PyYAML>=6.0.1          # Config file loading
tqdm>=4.66.1           # Progress bars
psutil>=5.9.0          # Memory monitoring
requests>=2.31.0       # HTTP requests in testing
python-dotenv>=1.0.0   # Environment management

# Optional Development Tools
tensorboard>=2.11.0    # Training visualization (optional)